#2: * adds single components for database structure list items
